clinic/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1
gunicorn==23.0.0
idna==3.8
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
djangorestframework==3.12.4
requests==2.26.0